fix: resolve 82 code review findings across all components
Safety-critical (Arduino):
- Reject NaN/Inf velocities in serial parser to prevent limit switch bypass
- Fix micros() wrap-around that stalled steppers after ~70 minutes
- Fix buffer size mismatch in serial protocol header
- Clamp step intervals to prevent undefined behavior on tiny velocities
- Add release debounce on limit switches, guarantee stepper pulse width

Security (Jetson):
- Enable TLS certificate verification on MQTT (configurable)
- Add velocity bounds and NaN rejection on MQTT command handler
- Replace os.system("sudo reboot") with subprocess.run
- Use hmac.compare_digest for timing-safe password comparison
- Add connection limits, security headers, and waitress WSGI server
- Warn on default HUD password, make bind address configurable

Architecture (Jetson):
- Consolidate PID ownership into SentryBrain (sole authority)
- Remove duplicate PID controllers from TurretManager
- Fix double heading offset in GPS computation (C-03)
- Fix LRF distance_m falsiness rejecting valid 0.0 readings
- Replace greedy centroid matching with Hungarian algorithm
- Select tracking target by threat score instead of bounding box area

Robustness (Jetson):
- Fix deadlock in ArduinoLink._reconnect (Lock -> RLock)
- Add send_enable() locking to prevent serial byte interleaving
- Protect all subsystem init with boot-failure counter
- Add main loop exception handling with consecutive error limit
- Fix MQTT client leak on reconnection, drain busy-loop
- Add heartbeat grace period, camera thread join on stop
- Add MVP software motion backstop using hard limits
- Replace deprecated datetime.utcnow() project-wide
- Switch to time.monotonic() in PID, add derivative low-pass filter
- Migrate to paho-mqtt v2 API

Flutter app:
- Make MJPEG video scheme configurable (HTTPS by default)
- Add dead-man's switch timeout on override joystick
- Fix silent background monitoring failure
- Add sentry_id validation on incoming MQTT messages
- Fix sessionId.substring crash on short IDs
- Fix copyWith nullable fields, cache refs in dispose

Infrastructure:
- Remove Docker privileged mode, add non-root user
- Pin Docker base image tag
- Remove hardcoded HUD credentials from docker-compose
- Pin requirements versions, split dev dependencies
- Fix MQTT port mismatch (1883 -> 8883)

+++ arduino/sentry_turret/src/stepper.h
@@ -31,8 +31,8 @@ struct StepperAxis {
     uint8_t       stepPin;           ///< Arduino digital pin number for STEP signal
     uint8_t       dirPin;            ///< Arduino digital pin number for DIR signal
     float         velocity;          ///< Current velocity in caller-defined speed units (negative = reverse)
-    unsigned long stepIntervalUs;    ///< Time between steps in microseconds (derived from velocity)
-    unsigned long nextStepTimeUs;    ///< micros() timestamp of the next scheduled step
+    uint32_t      stepIntervalUs;    ///< Time between steps in microseconds (derived from velocity)
+    uint32_t      nextStepTimeUs;    ///< micros() timestamp of the next scheduled step
     int32_t       stepCount;         ///< Cumulative signed step count (positive = forward)
 };
